+++ monitor-alert/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Release v1.2.0
+
+## Provider & Terraform Upgrade
+- Azurerm provider: 3.73.0 (#6 )
+- Terraform: 1.5.7 (#6)
+
+## Enhancements
+
+- Add azurerm_monitor_metric_alert resource (#5)
+
+
+   
 ## Release v1.1.0
 
 ### Version upgrade